Pull more patching logic out, this time out of linted_file
