+++ src/sqlfluff/core/linter/linted_file.py
@@ -21,7 +21,7 @@ from sqlfluff.core.errors import (
     SQLParseError,
     SQLTemplaterError,
 )
-from sqlfluff.core.linter.patch import FixPatch, iter_patches
+from sqlfluff.core.linter.patch import FixPatch, generate_source_patches
 
 # Classes needed only for type checking
 from sqlfluff.core.parser.segments import BaseSegment
@@ -177,30 +177,6 @@ class LintedFile(NamedTuple):
         """Return True if there are no ignorable violations."""
         return not any(self.get_violations(filter_ignore=True))
 
-    @staticmethod
-    def _log_hints(patch: FixPatch, templated_file: TemplatedFile):
-        """Log hints for debugging during patch generation."""
-        # This next bit is ALL FOR LOGGING AND DEBUGGING
-        max_log_length = 10
-        if patch.templated_slice.start >= max_log_length:
-            pre_hint = templated_file.templated_str[
-                patch.templated_slice.start
-                - max_log_length : patch.templated_slice.start
-            ]
-        else:
-            pre_hint = templated_file.templated_str[: patch.templated_slice.start]
-        if patch.templated_slice.stop + max_log_length < len(
-            templated_file.templated_str
-        ):
-            post_hint = templated_file.templated_str[
-                patch.templated_slice.stop : patch.templated_slice.stop + max_log_length
-            ]
-        else:
-            post_hint = templated_file.templated_str[patch.templated_slice.stop :]
-        linter_logger.debug(
-            "        Templated Hint: ...%r <> %r...", pre_hint, post_hint
-        )
-
     def fix_string(self) -> Tuple[str, bool]:
         """Obtain the changes to a path as a string.
 
@@ -243,7 +219,7 @@ class LintedFile(NamedTuple):
         # Generate patches from the fixed tree. In the process we sort
         # and deduplicate them so that the resultant list is in the
         # the right order for the source file without any duplicates.
-        filtered_source_patches = self._generate_source_patches(
+        filtered_source_patches = generate_source_patches(
             self.tree, self.templated_file
         )
         linter_logger.debug("Filtered source patches:")
@@ -270,89 +246,6 @@ class LintedFile(NamedTuple):
 
         # The success metric here is whether anything ACTUALLY changed.
         return fixed_source_string, fixed_source_string != original_source
-
-    @classmethod
-    def _generate_source_patches(
-        cls, tree: BaseSegment, templated_file: TemplatedFile
-    ) -> List[FixPatch]:
-        """Use the fixed tree to generate source patches.
-
-        Importantly here we deduplicate and sort the patches
-        from their position in the templated file into their
-        intended order in the source file.
-        """
-        # Iterate patches, filtering and translating as we go:
-        linter_logger.debug("### Beginning Patch Iteration.")
-        filtered_source_patches = []
-        dedupe_buffer = []
-        # We use enumerate so that we get an index for each patch. This is entirely
-        # so when debugging logs we can find a given patch again!
-        for idx, patch in enumerate(iter_patches(tree, templated_file=templated_file)):
-            linter_logger.debug("  %s Yielded patch: %s", idx, patch)
-            cls._log_hints(patch, templated_file)
-
-            # Check for duplicates
-            if patch.dedupe_tuple() in dedupe_buffer:
-                linter_logger.info(
-                    "      - Skipping. Source space Duplicate: %s",
-                    patch.dedupe_tuple(),
-                )
-                continue
-
-            # We now evaluate patches in the source-space for whether they overlap
-            # or disrupt any templated sections.
-            # The intent here is that unless explicitly stated, a fix should never
-            # disrupt a templated section.
-            # NOTE: We rely here on the patches being generated in order.
-            # TODO: Implement a mechanism for doing templated section fixes. Given
-            # these patches are currently generated from fixed segments, there will
-            # likely need to be an entirely different mechanism
-
-            # Get the affected raw slices.
-            local_raw_slices = templated_file.raw_slices_spanning_source_slice(
-                patch.source_slice
-            )
-            local_type_list = [slc.slice_type for slc in local_raw_slices]
-
-            # Deal with the easy cases of 1) New code at end 2) only literals
-            if not local_type_list or set(local_type_list) == {"literal"}:
-                linter_logger.info(
-                    "      * Keeping patch on new or literal-only section.",
-                )
-                filtered_source_patches.append(patch)
-                dedupe_buffer.append(patch.dedupe_tuple())
-            # Handle the easy case of an explicit source fix
-            elif patch.patch_category == "source":
-                linter_logger.info(
-                    "      * Keeping explicit source fix patch.",
-                )
-                filtered_source_patches.append(patch)
-                dedupe_buffer.append(patch.dedupe_tuple())
-            # Is it a zero length patch.
-            elif (
-                patch.source_slice.start == patch.source_slice.stop
-                and patch.source_slice.start == local_raw_slices[0].source_idx
-            ):
-                linter_logger.info(
-                    "      * Keeping insertion patch on slice boundary.",
-                )
-                filtered_source_patches.append(patch)
-                dedupe_buffer.append(patch.dedupe_tuple())
-            else:  # pragma: no cover
-                # We've got a situation where the ends of our patch need to be
-                # more carefully mapped. This used to happen with greedy template
-                # element matching, but should now never happen. In the event that
-                # it does, we'll warn but carry on.
-                linter_logger.warning(
-                    "Skipping edit patch on uncertain templated section [%s], "
-                    "Please report this warning on GitHub along with the query "
-                    "that produced it.",
-                    (patch.patch_category, patch.source_slice),
-                )
-                continue
-
-        # Sort the patches before building up the file.
-        return sorted(filtered_source_patches, key=lambda x: x.source_slice.start)
 
     @staticmethod
     def _slice_source_file_using_patches(
