+++ src/sqlfluff/core/linter/patch.py
@@ -4,6 +4,7 @@ import logging
 from dataclasses import dataclass
 from typing import (
     Iterator,
+    List,
     Tuple,
 )
 
@@ -54,7 +55,7 @@ def _iter_source_fix_patches(
         )
 
 
-def iter_patches(
+def _iter_templated_patches(
     segment: BaseSegment, templated_file: TemplatedFile
 ) -> Iterator[FixPatch]:
     """Iterate through the segments generating fix patches.
@@ -175,7 +176,7 @@ def iter_patches(
                 insert_buff = ""
 
             # Now we deal with any changes *within* the segment itself.
-            yield from iter_patches(seg, templated_file=templated_file)
+            yield from _iter_templated_patches(seg, templated_file=templated_file)
 
             # Once we've dealt with any patches from the segment, update
             # our position markers.
@@ -207,3 +208,103 @@ def iter_patches(
                 templated_str=templated_file.templated_str[templated_slice],
                 source_str=templated_file.source_str[source_slice],
             )
+
+
+def _log_hints(patch: FixPatch, templated_file: TemplatedFile):
+    """Log hints for debugging during patch generation."""
+    max_log_length = 10
+    if patch.templated_slice.start >= max_log_length:
+        pre_hint = templated_file.templated_str[
+            patch.templated_slice.start - max_log_length : patch.templated_slice.start
+        ]
+    else:
+        pre_hint = templated_file.templated_str[: patch.templated_slice.start]
+    if patch.templated_slice.stop + max_log_length < len(templated_file.templated_str):
+        post_hint = templated_file.templated_str[
+            patch.templated_slice.stop : patch.templated_slice.stop + max_log_length
+        ]
+    else:
+        post_hint = templated_file.templated_str[patch.templated_slice.stop :]
+    linter_logger.debug("        Templated Hint: ...%r <> %r...", pre_hint, post_hint)
+
+
+def generate_source_patches(
+    tree: BaseSegment, templated_file: TemplatedFile
+) -> List[FixPatch]:
+    """Use the fixed tree to generate source patches.
+
+    Importantly here we deduplicate and sort the patches from their position
+    in the templated file into their intended order in the source file.
+
+    Any source fixes are generated in `_iter_templated_patches` and included
+    alongside any standard fixes. That means we treat them the same here.
+    """
+    # Iterate patches, filtering and translating as we go:
+    linter_logger.debug("### Beginning Patch Iteration.")
+    filtered_source_patches = []
+    dedupe_buffer = []
+    # We use enumerate so that we get an index for each patch. This is entirely
+    # so when debugging logs we can find a given patch again!
+    for idx, patch in enumerate(
+        _iter_templated_patches(tree, templated_file=templated_file)
+    ):
+        linter_logger.debug("  %s Yielded patch: %s", idx, patch)
+        _log_hints(patch, templated_file)
+
+        # Check for duplicates
+        if patch.dedupe_tuple() in dedupe_buffer:
+            linter_logger.info(
+                "      - Skipping. Source space Duplicate: %s",
+                patch.dedupe_tuple(),
+            )
+            continue
+
+        # We now evaluate patches in the source-space for whether they overlap
+        # or disrupt any templated sections unless designed to do so.
+        # NOTE: We rely here on the patches being generated in order.
+
+        # Get the affected raw slices.
+        local_raw_slices = templated_file.raw_slices_spanning_source_slice(
+            patch.source_slice
+        )
+        local_type_list = [slc.slice_type for slc in local_raw_slices]
+
+        # Deal with the easy cases of 1) New code at end 2) only literals
+        if not local_type_list or set(local_type_list) == {"literal"}:
+            linter_logger.info(
+                "      * Keeping patch on new or literal-only section.",
+            )
+            filtered_source_patches.append(patch)
+            dedupe_buffer.append(patch.dedupe_tuple())
+        # Handle the easy case of an explicit source fix
+        elif patch.patch_category == "source":
+            linter_logger.info(
+                "      * Keeping explicit source fix patch.",
+            )
+            filtered_source_patches.append(patch)
+            dedupe_buffer.append(patch.dedupe_tuple())
+        # Is it a zero length patch.
+        elif (
+            patch.source_slice.start == patch.source_slice.stop
+            and patch.source_slice.start == local_raw_slices[0].source_idx
+        ):
+            linter_logger.info(
+                "      * Keeping insertion patch on slice boundary.",
+            )
+            filtered_source_patches.append(patch)
+            dedupe_buffer.append(patch.dedupe_tuple())
+        else:  # pragma: no cover
+            # We've got a situation where the ends of our patch need to be
+            # more carefully mapped. This used to happen with greedy template
+            # element matching, but should now never happen. In the event that
+            # it does, we'll warn but carry on.
+            linter_logger.warning(
+                "Skipping edit patch on uncertain templated section [%s], "
+                "Please report this warning on GitHub along with the query "
+                "that produced it.",
+                (patch.patch_category, patch.source_slice),
+            )
+            continue
+
+    # Sort the patches before building up the file.
+    return sorted(filtered_source_patches, key=lambda x: x.source_slice.start)
