+++ test/core/linter/linted_file_test.py
@@ -6,15 +6,7 @@ import pytest
 
 from sqlfluff.core.linter import LintedFile
 from sqlfluff.core.linter.patch import FixPatch
-from sqlfluff.core.parser.markers import PositionMarker
-from sqlfluff.core.parser.segments import (
-    BaseSegment,
-    RawSegment,
-    TemplateSegment,
-)
-from sqlfluff.core.parser.segments.raw import SourceFix
-from sqlfluff.core.templaters import RawFileSlice, TemplatedFile
-from sqlfluff.core.templaters.base import TemplatedFileSlice
+from sqlfluff.core.templaters import RawFileSlice
 
 
 @pytest.mark.parametrize(
@@ -215,157 +207,6 @@ def test__linted_file__slice_source_file_using_patches(
         result = LintedFile._slice_source_file_using_patches(
             source_patches, source_only_slices, raw_source_string
         )
-    assert result == expected_result
-
-
-templated_file_1 = TemplatedFile.from_string("abc")
-templated_file_2 = TemplatedFile(
-    "{# blah #}{{ foo }}bc",
-    "<testing>",
-    "abc",
-    [
-        TemplatedFileSlice("comment", slice(0, 10), slice(0, 0)),
-        TemplatedFileSlice("templated", slice(10, 19), slice(0, 1)),
-        TemplatedFileSlice("literal", slice(19, 21), slice(1, 3)),
-    ],
-    [
-        RawFileSlice("{# blah #}", "comment", 0),
-        RawFileSlice("{{ foo }}", "templated", 10),
-        RawFileSlice("bc", "literal", 19),
-    ],
-)
-
-
-@pytest.mark.parametrize(
-    "tree,templated_file,expected_result",
-    [
-        # Trivial example
-        (
-            RawSegment(
-                "abc",
-                PositionMarker(slice(0, 3), slice(0, 3), templated_file_1),
-                "code",
-            ),
-            templated_file_1,
-            [],
-        ),
-        # Simple literal edit example
-        (
-            RawSegment(
-                "abz",
-                PositionMarker(slice(0, 3), slice(0, 3), templated_file_1),
-                "code",
-            ),
-            templated_file_1,
-            [FixPatch(slice(0, 3), "abz", "literal", slice(0, 3), "abc", "abc")],
-        ),
-        # Nested literal edit example
-        (
-            BaseSegment(
-                [
-                    RawSegment(
-                        "a",
-                        PositionMarker(slice(0, 1), slice(0, 1), templated_file_1),
-                        "code",
-                    ),
-                    RawSegment(
-                        "b",
-                        PositionMarker(slice(1, 2), slice(1, 2), templated_file_1),
-                        "code",
-                    ),
-                    RawSegment(
-                        "z",
-                        PositionMarker(slice(2, 3), slice(2, 3), templated_file_1),
-                        "code",
-                    ),
-                ]
-            ),
-            templated_file_1,
-            [FixPatch(slice(0, 3), "abz", "literal", slice(0, 3), "abc", "abc")],
-        ),
-        # More complicated templating example
-        (
-            BaseSegment(
-                [
-                    TemplateSegment(
-                        PositionMarker(slice(0, 10), slice(0, 0), templated_file_2),
-                        "{# blah #}",
-                        "comment",
-                    ),
-                    RawSegment(
-                        "a",
-                        PositionMarker(slice(10, 20), slice(0, 1), templated_file_2),
-                        "code",
-                    ),
-                    RawSegment(
-                        "b",
-                        PositionMarker(slice(19, 20), slice(1, 2), templated_file_2),
-                        "code",
-                    ),
-                    RawSegment(
-                        "z",
-                        PositionMarker(slice(20, 21), slice(2, 3), templated_file_2),
-                        "code",
-                    ),
-                ]
-            ),
-            templated_file_2,
-            [FixPatch(slice(2, 3), "z", "literal", slice(20, 21), "c", "c")],
-        ),
-        # Templating example with fixes
-        (
-            BaseSegment(
-                [
-                    TemplateSegment(
-                        PositionMarker(slice(0, 10), slice(0, 0), templated_file_2),
-                        "{# blah #}",
-                        "comment",
-                        source_fixes=[
-                            SourceFix("{# fixed #}", slice(0, 10), slice(0, 0))
-                        ],
-                    ),
-                    RawSegment(
-                        "a",
-                        PositionMarker(slice(10, 19), slice(0, 1), templated_file_2),
-                        "code",
-                        source_fixes=[
-                            SourceFix("{{ bar }}", slice(10, 19), slice(0, 1))
-                        ],
-                    ),
-                    RawSegment(
-                        "b",
-                        PositionMarker(slice(19, 20), slice(1, 2), templated_file_2),
-                        "code",
-                    ),
-                    RawSegment(
-                        "z",
-                        PositionMarker(slice(20, 21), slice(2, 3), templated_file_2),
-                        "code",
-                    ),
-                ]
-            ),
-            templated_file_2,
-            [
-                FixPatch(
-                    slice(0, 0), "{# fixed #}", "source", slice(0, 10), "", "{# blah #}"
-                ),
-                FixPatch(
-                    slice(0, 1), "{{ bar }}", "source", slice(10, 19), "a", "{{ foo }}"
-                ),
-                FixPatch(slice(2, 3), "z", "literal", slice(20, 21), "c", "c"),
-            ],
-        ),
-    ],
-)
-def test__linted_file__generate_source_patches(
-    tree, templated_file, expected_result, caplog
-):
-    """Test _generate_source_patches.
-
-    This is part of fix_string().
-    """
-    with caplog.at_level(logging.DEBUG, logger="sqlfluff.linter"):
-        result = LintedFile._generate_source_patches(tree, templated_file)
     assert result == expected_result
