+++ test/fixtures/rules/std_rule_cases/CP03.yml
@@ -77,7 +77,7 @@ test_fail_ignore_templated_code_false:
       ignore_templated_areas: false
 
 test_pass_func_name_templated_literal_mix:
-  # Issue 3022. This was actually a bug in linter.patch.iter_patches().
+  # Issue 3022. This was actually a bug in linter.patch._iter_templated_patches().
   pass_str: SELECT RO(), {{ "t" }}.func()
 
 test_pass_ignore_words_regex_simple:
